+++ oci/health_checks_http_monitor_data_source.go
@@ -78,6 +78,10 @@ func (s *HealthChecksHttpMonitorDataSourceCrud) SetData() error {
 
 	s.D.Set("headers", s.Res.Headers)
 
+	if s.Res.HomeRegion != nil {
+		s.D.Set("home_region", *s.Res.HomeRegion)
+	}
+
 	if s.Res.IntervalInSeconds != nil {
 		s.D.Set("interval_in_seconds", *s.Res.IntervalInSeconds)
 	}
@@ -103,6 +107,10 @@ func (s *HealthChecksHttpMonitorDataSourceCrud) SetData() error {
 	}
 
 	s.D.Set("targets", s.Res.Targets)
+
+	if s.Res.TimeCreated != nil {
+		s.D.Set("time_created", s.Res.TimeCreated.String())
+	}
 
 	if s.Res.TimeoutInSeconds != nil {
 		s.D.Set("timeout_in_seconds", *s.Res.TimeoutInSeconds)